Add files via upload
The aim is to provide you a refresher on the Pandas library, so that you can pre-process and anlyse the datasets before applying data visualization techniques on it. This will work as acrash course on *pandas*. if you are interested in learning more about the *pandas* library, detailed description and explanation of how to use it and how to clean, munge, and process data stored in a *pandas* dataframe you can check this repository. In this file you will get:
1. Exploring Datasets with pandas
2. The Dataset: Immigration to Canada from 1980 to 2013
3. Pandas Basics
4. Pandas Intermediate: Indexing and Selection
5. Pandas Filtering based on a criteria
6. Pandas Sorting Values